Add native for deleting a module in module manager.
The module manager doesn't know when a plugin/module is unloaded, so the module has to notifiy the module manager. Without notification, the module manager will leak resources and won't be able to notify other modules.

The optimal solution is that the module manager monitor plugins itself, but that's not possible in SourceMod at the moment.